chore: 同步更新权重类型为 double

+++ RP4CI.Interface/Models/PickGroup.cs
@@ -6,6 +6,6 @@ public partial class PickGroup : ObservableRecipient
 {
     [ObservableProperty] private string _name = string.Empty;
     [ObservableProperty] private List<PickStudent>? _studentList;
-    [ObservableProperty] private int _weight = 1;
+    [ObservableProperty] private double _weight = 1.0;
     [ObservableProperty] private bool _active = true;
 }
