+++ RP4CI.Interface/Models/PickStudent.cs
@@ -6,6 +6,6 @@ public partial class PickStudent : ObservableRecipient
 {
     [ObservableProperty] private string _name = string.Empty;
     [ObservableProperty] private List<CustomProperty>? _customProperties;
-    [ObservableProperty] private int _weight = 1;
+    [ObservableProperty] private double _weight = 1.0;
     [ObservableProperty] private bool _active = true;
 }
